+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 1.0.5
+Remove default props for lat,lng and radius. 
+Add queryCountries where you can pass array of countries to limit your search results. 
 ## 1.0.4
 Update changelog and readme
 ## 1.0.3